Move map conditions to dedicated file
